% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfunctions.R
\name{rnaseq_df_to_matrix_gene_columns}
\alias{rnaseq_df_to_matrix_gene_columns}
\title{Expression Dataframe to Matrix}
\usage{
rnaseq_df_to_matrix_gene_columns(rnaseq_df)
}
\arguments{
\item{rnaseq_df}{an rnaseq dataframe read by read_rnaseq_file}
}
\value{
matrix where columns are genes and rows are samples
}
\description{
Expression Dataframe to Matrix
}
\examples{
path=system.file("testfiles/blca_rnaseq.tsv", package="utilitybeltrna")
rnaseq_df_to_matrix_gene_columns(read_rnaseq_data(path))

}
