% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfunctions.R
\name{rnaseq_df_to_matrix_tsb_columns}
\alias{rnaseq_df_to_matrix_tsb_columns}
\title{RNAseq_df to matrix}
\usage{
rnaseq_df_to_matrix_tsb_columns(rnaseq_df)
}
\arguments{
\item{rnaseq_df}{an rnaseq dataframe read by read_rnaseq_file}
}
\value{
a matrix where rows are genes and columns are samples. Each value represents TPM of a gene in a single sample (matrix)
}
\description{
Converts RNAseq dataframe to a matrix that will be used as the input for most downstream
}
\examples{
path=system.file("testfiles/blca_rnaseq.tsv", package="utilitybeltrna")
rnaseq_df_to_matrix_tsb_columns(read_rnaseq_data(path))
}
