% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfunctions.R
\name{rnaseq_get_metadata_for_plotting}
\alias{rnaseq_get_metadata_for_plotting}
\title{Get metadata for plotting}
\usage{
rnaseq_get_metadata_for_plotting(
  rnaseq_tsne_matrix,
  metadata_df,
  column_containing_sample_ids = "Tumor_Sample_Barcode",
  feature_of_interest
)
}
\arguments{
\item{rnaseq_tsne_matrix}{the gene expression matrix supplied to rnaseq_tsne}

\item{metadata_df}{a dataframe containing clinical metadata. Must contain a 'Tumor_Sample_Barcode' column with unique sampleIDs}

\item{feature_of_interest}{the name of the column in metadata_df you want to annotate in a TSNE plot}
}
\value{
a vector of 'feature_of_interest' for use with rnaseq_tsne_ggplot
}
\description{
Get metadata for plotting
}
\examples{
# Transpose GDC data imported using GDCRNATools so sampleids are on rows
rnaseq_mx <- t(data_expression_matrix_gbm_gdc)

# Run tsne
tsne_out <- rnaseq_tsne(rnaseq_mx)

# Get some metadata of interest to color points by
vital_status <- rnaseq_get_metadata_for_plotting(
 rnaseq_tsne_matrix = rnaseq_mx,
 metadata_df = data_metadata_gbm_gdc,
 column_containing_sample_ids = "sample",
 feature_of_interest = "vital_status"
)

# Plot Rtsne results
    rnaseq_tsne_output = tsne_out,
    feature_name_color = "Vital_Status",
    feature_values_color = vital_status
)
}
