% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfunctions.R
\name{rnaseq_tsne}
\alias{rnaseq_tsne}
\title{tsne}
\usage{
rnaseq_tsne(rnaseq_matrix, perplexity = 30)
}
\arguments{
\item{rnaseq_matrix}{an rnaseq data matrix where genes are columns, samples are rows and values are TPM or some other metric of expression}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}
}
\value{
tsne results in a list-format. Plot results$Y to see (list)
}
\description{
Perform dimensionality reduction on a tsne dataset
}
\examples{
\dontrun{
# Transpose GDC data imported using GDCRNATools so sampleids are on rows
rnaseq_mx <- t(data_expression_matrix_gbm_gdc)

# Run tsne
tsne_out <- rnaseq_tsne(rnaseq_mx)

# Plot Rtsne results
 rnaseq_tsne_ggplot(
     rnaseq_tsne_output = tsne_out,
     sample_metadata_df = data_metadata_gbm_gdc,
     color_column = "vital_status",
     sample_column = "sample"
 )
 }
}
