% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfunctions.R
\name{rnaseq_tsne_ggplot}
\alias{rnaseq_tsne_ggplot}
\title{rnaseq_tsne_ggplot}
\usage{
rnaseq_tsne_ggplot(
  rnaseq_tsne_output,
  sample_metadata_df,
  color_column = NULL,
  sample_column = "Tumor_Sample_Barcode"
)
}
\arguments{
\item{rnaseq_tsne_output}{output from utilitybeltrna::rnaseq_tsne (list)}

\item{color_column}{name of sample metadata column to base point color on (string)}

\item{sample_column}{name of sample metadata column containing sample ids (string)}
}
\value{
ggplot object
}
\description{
rnaseq_tsne_ggplot
}
\examples{
\dontrun{
# Transpose GDC data imported using GDCRNATools so sampleids are on rows
rnaseq_mx <- t(data_expression_matrix_gbm_gdc)

# Run tsne
tsne_out <- rnaseq_tsne(rnaseq_mx)

# Plot Rtsne results
 rnaseq_tsne_ggplot(
     rnaseq_tsne_output = tsne_out,
     sample_metadata_df = data_metadata_gbm_gdc,
     color_column = "vital_status",
     sample_column = "sample"
 )
 }
}
